Restart instead of reload.
The init script does not seem to support reloading, when called the
process is terminated and never restarted - use restart instead.

Also, moved away from using handlers for restarting after configuration
changes because handlers are not properly flushed between roles in a
play. So if one role in a play starts mongod and another role in the
play relies on it, problems arise as the handlers are flushed between
plays and there is no convenient way to call meta: flush_handlers
between roles.